Sorry for single commit.
I had to do this work ASAP, so issues presence is obvious. Despite the
fact that this app works and I tried to keep all the objective clauses,
there are still shortcomings, such as lack of the system resources
release routines, unsafe usage of the shared resources by threads and
several less significant problems.